Se agregan dependencias al pom.xml, se organizan los modelos y se elimina modelo obsoleto del caso
