+++ src/main/java/com/sumativa/b/consultorio/model/AtencionMedica.java
@@ -1,0 +1,90 @@
+package com.sumativa.b.consultorio.model;
+
+import java.util.Date;
+
+import com.fasterxml.jackson.annotation.JsonBackReference;
+import jakarta.persistence.Column;
+import jakarta.persistence.Entity;
+import jakarta.persistence.FetchType;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.GenerationType;
+import jakarta.persistence.Id;
+import jakarta.persistence.JoinColumn;
+import jakarta.persistence.ManyToOne;
+import jakarta.persistence.Table;
+
+@Entity
+@Table(name = "Atencion_Medica")
+public class AtencionMedica {
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    @Column(name = "id_atencion")
+    private Long id;
+    @Column(name = "fecha_atencion")
+    private Date fecha;
+    @Column(name = "tipo_atencion")
+    private String tipo; 
+    @Column(name = "descripcion_atencion")
+    private String descripcion;
+    @ManyToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name = "id_paciente")
+    @JsonBackReference
+    private Paciente paciente;
+
+    public AtencionMedica() {
+    }
+
+    public AtencionMedica(Long id, Date fecha, String tipo, String descripcion, Paciente paciente) {
+        this.id = id;
+        this.fecha = fecha;
+        this.tipo = tipo;
+        this.descripcion = descripcion;
+        this.paciente = paciente;
+    }
+
+    public Long getId() {
+        return id;
+    }
+
+    public void setId(Long id) {
+        this.id = id;
+    }
+
+    public Date getFecha() {
+        return fecha;
+    }
+
+    public void setFecha(Date fecha) {
+        this.fecha = fecha;
+    }
+
+    public String getTipo() {
+        return tipo;
+    }
+
+    public void setTipo(String tipo) {
+        this.tipo = tipo;
+    }
+
+    public String getDescripcion() {
+        return descripcion;
+    }
+
+    public void setDescripcion(String descripcion) {
+        this.descripcion = descripcion;
+    }
+
+    public Paciente getPaciente() {
+        return paciente;
+    }
+
+    public void setPaciente(Paciente paciente) {
+        this.paciente = paciente;
+    }
+
+    
+   
+    
+
+    
+}
