+++ src/main/java/com/sumativa/b/consultorio/model/ConsultaMedica.java
@@ -1,0 +1,100 @@
+package com.sumativa.b.consultorio.model;
+
+import java.util.Date;
+
+import com.fasterxml.jackson.annotation.JsonBackReference;
+
+import jakarta.persistence.Column;
+import jakarta.persistence.Entity;
+import jakarta.persistence.FetchType;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.GenerationType;
+import jakarta.persistence.Id;
+import jakarta.persistence.JoinColumn;
+import jakarta.persistence.ManyToOne;
+import jakarta.persistence.Table;
+
+@Entity
+@Table(name = "Consulta_Medica")
+public class ConsultaMedica {
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    @Column(name = "id_consulta")
+    private Long id;
+    @Column(name = "fecha_consulta")
+    private Date fecha;
+    @Column(name = "motivo_consulta")
+    private String motivoConsulta;
+    @Column(name = "diganostico_consulta")
+    private String diagnostico;
+    @Column(name = "tratamiento_consulta")
+    private String tratamiento;
+    @ManyToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name = "id_paciente")
+    @JsonBackReference
+    private Paciente paciente;
+    
+    public ConsultaMedica() {
+    }
+
+    public ConsultaMedica(Long id, Date fecha, String motivoConsulta, String diagnostico, String tratamiento,
+            Paciente paciente) {
+        this.id = id;
+        this.fecha = fecha;
+        this.motivoConsulta = motivoConsulta;
+        this.diagnostico = diagnostico;
+        this.tratamiento = tratamiento;
+        this.paciente = paciente;
+    }
+
+    public Long getId() {
+        return id;
+    }
+
+    public void setId(Long id) {
+        this.id = id;
+    }
+
+    public Date getFecha() {
+        return fecha;
+    }
+
+    public void setFecha(Date fecha) {
+        this.fecha = fecha;
+    }
+
+    public String getMotivoConsulta() {
+        return motivoConsulta;
+    }
+
+    public void setMotivoConsulta(String motivoConsulta) {
+        this.motivoConsulta = motivoConsulta;
+    }
+
+    public String getDiagnostico() {
+        return diagnostico;
+    }
+
+    public void setDiagnostico(String diagnostico) {
+        this.diagnostico = diagnostico;
+    }
+
+    public String getTratamiento() {
+        return tratamiento;
+    }
+
+    public void setTratamiento(String tratamiento) {
+        this.tratamiento = tratamiento;
+    }
+
+    public Paciente getPaciente() {
+        return paciente;
+    }
+
+    public void setPaciente(Paciente paciente) {
+        this.paciente = paciente;
+    }
+
+    
+    
+}
